requirements.txt
fastapi
uvicorn
tensorflow==2.15.0
pillow
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiofiles